DELIA-38481 : address OSX compiler warnings while building aamp-cli
Reason for change : Fixed the OSX compiler warnings, also wrote script for patching and installing libdash
Test Procedure : tested on macOSX
Risks : None



DELIA-38481: [AAMP] Fix compiler warnings

Reason for change: Fixed GCC compiler warnings reported in RDK Yocto builds
Test Procedure: tested on arris xg1v4
Risks: None

+++ test/aampcli.cpp
@@ -356,7 +356,7 @@ static void ProcessCliCommand(char *cmd)
 	}
 	else if (strcmp(cmd, "sf") == 0)
 	{
-		mSingleton->SetRate(0.5);
+		mSingleton->SetRate((int)0.5);
 	}
 	else if (sscanf(cmd, "ff%d", &rate) == 1)
 	{
